example/model_serving/tensorflow/requirements.txt
opencv-python
tensorflow==1.14
tensorflow-serving-api==1.14
numpy
grpcio==1.25 --no-binary grpcio
grpcio-tools==1.25
